+++ packages/themes/builtin/xianxia/src/cultivation/index.ts
@@ -1,0 +1,11 @@
+export { computeCultivation, isSameCultivationSnapshot } from "./computeCultivation";
+export { CULTIVATION_REALMS } from "./realms";
+export { computeCultivationScore } from "./score";
+export {
+	CULTIVATION_SNAPSHOT_VERSION,
+	CULTIVATION_STORAGE_KEY,
+	type CultivationRealmDefinition,
+	type CultivationScoreBreakdown,
+	type CultivationSnapshot,
+} from "./types";
+export { XianxiaCultivationRuntime } from "./XianxiaCultivationRuntime";